Refactor legendary crafting scripts
